Use a single dataview to show domain connection setup tables
